feat(compass-global-writes): sharding error state
